fix bugs

+++ src/main/java/villages/dao/VilleDao.java
@@ -3,6 +3,6 @@ package villages.dao;
 import villages.dao.Dao;
 import villages.model.Ville;
 
-public interface VilleDao extends Dao<Ville, Integer> {
+public interface VilleDao extends Dao<Ville, Long> {
 
 }
